Use new criteria for clean_git and update_git
clean_git: remove file if exist
update_git: increase window width and depth for repack

+++ update_git.sh
@@ -65,7 +65,9 @@ loop() {
 
 start_folder="$(pwd)"
 
-rm "${abandonned_file}"
+if [ -f "${abandonned_file}" ]; then
+	rm "${abandonned_file}"
+fi
 if [[ "$1" == "." ]]; then
 	builtin cd "${start_folder}"
 	loop "."
